separate test for easier readability and scope
