+++ tests/Rules/UnusedPublicClassMethodRule/templates/blade/used_of_static_call.blade.php
@@ -1,0 +1,5 @@
+<div>
+
+{{ \TomasVotruba\UnusedPublic\Tests\Rules\UnusedPublicClassMethodRule\Fixture\Blade\SkipStaticCallUsedInBlade::countMeOut() }}
+
+</div>
